added EditAndResendDocumentModal,  changed documentsPage and tableComponent
