+++ fixtures/base.js
@@ -17,6 +17,7 @@ import SpecialOneTimeOfferModal from "../new_pom/modalWindows/specialOneTimeOffe
 import CreateSignatureModal from "../new_pom/modalWindows/createSignatureModal.js";
 import FinalStepPage from '../new_pom/pages/finalStepPage.js';
 import SuccessModal from '../new_pom/modalWindows/successModal.js';
+import EditAndResendDocumentModal from "../new_pom/modalWindows/editAndResendDocumentModal.js";
 
 
 export const test = base.extend({    
@@ -113,6 +114,10 @@ export const test = base.extend({
 
     successModal: async ({ page }, use) => {
         await use(new SuccessModal(page));
+    },
+
+    editAndResendDocumentModal: async ({ page }, use) => {
+        await use(new EditAndResendDocumentModal(page));
     },
 
 });
