+++ new_pom/components/tableComponent.js
@@ -5,6 +5,17 @@ export default class TableComponent {
 
         this.emptyTableHeader = this.page.locator('.empty-table__header');
         this.documentStatus = this.page.locator('.documents__documentStatus').first();
+        this.optionsBtn = this.page.getByText('Options');
+        this.editAndResendBtn = this.page.getByText('Edit & Resend');
+    }
+
+    async clickOptionsBtn() {
+        await this.optionsBtn.waitFor();
+        await this.optionsBtn.click();
+    }
+
+    async clickEditAndResendBtn() {
+        await this.editAndResendBtn.click();
     }
 
 }
