+++ new_pom/modalWindows/editAndResendDocumentModal.js
@@ -1,0 +1,8 @@
+export default class EditAndResendDocumentModal {
+  constructor(page) {
+      this.page = page;
+
+      this.editAndResendTitle = this.page.getByText('Edit & Resend document');
+  }
+
+}
